+++ include/joint_level_control/hall_sensor/swing_sensor_rosinterface.h
@@ -4,6 +4,6 @@
 #include <string>
 
 // reads angle from a swing sensor, with given ID
-double readSwingAngle(const std::string& spi_device, uint8_t spi_cs_id, uint8_t spi_mode, uint8_t spi_bits, uint8_t spi_speed, uint8_t spi_delay);
+double readSwingAngle(const std::string& spi_device, uint8_t spi_cs_id, uint8_t spi_mode, uint8_t spi_bits, uint32_t spi_speed, uint16_t spi_delay);
 
 #endif
