+++ src/joint/extend_joint.cpp
@@ -1,7 +1,7 @@
 #include "joint_level_control/joint/extend_joint.h"
 
 
-ExtendJoint::ExtendJoint(const std::string& joint_name, const std::string& spi_device, uint8_t spi_cs_id, uint8_t spi_mode, uint8_t spi_bits, uint8_t spi_speed, uint8_t spi_delay, const std::string& can_name, uint8_t can_id)
+ExtendJoint::ExtendJoint(const std::string& joint_name, const std::string& spi_device, uint8_t spi_cs_id, uint8_t spi_mode, uint8_t spi_bits, uint32_t spi_speed, uint16_t spi_delay, const std::string& can_name, uint8_t can_id)
     : position_(0.0), velocity_(0.0), effort_(0.0), command_position_(0.0), motor_(can_name, can_id), rotary_encoder_(spi_device, spi_cs_id, spi_mode, spi_bits, spi_speed, spi_delay)
 {
     // Setup hardware interface:
